Completed Tic Tac Toe Project
Gameboard is stored as an array called markers.
gameBoard.empty() is used to reset the gameboard.
gameBoard.display() is used to render the contents of markers to the
webpage.
placeMarker is used to check whose turn it is, then place the correct
marker.
At every click of a tile, checkWin is called.
Dialog box appears on site load in and when a game is completed.
Basic styles to make the webpage neater.